split more functions out, lowercase filenames
